+++ mobile/android/android-components/components/feature/top-sites/src/main/java/mozilla/components/feature/top/sites/TopSitesStorage.kt
@@ -5,7 +5,6 @@
 package mozilla.components.feature.top.sites
 
 import mozilla.components.browser.storage.sync.PlacesHistoryStorage
-import mozilla.components.concept.storage.FrecencyThresholdOption
 import mozilla.components.support.base.observer.Observable
 
 /**
@@ -43,15 +42,14 @@ interface TopSitesStorage : Observable<TopSitesStorage.Observer> {
      * If `frecencyConfig` is specified, fill in any missing top sites with frecent top site results.
      *
      * @param totalSites A total number of sites that will be retrieve if possible.
-     * @param frecencyConfig If [frecencyConfig] is specified, only visited sites with a frecency
-     * score above the given threshold will be returned. Otherwise, frecent top site results are
-     * not included.
+     * @param frecencyConfig An instance of [TopSitesFrecencyConfig] that specifies which top
+     * frecent sites to be included.
      * @param providerConfig An instance of [TopSitesProviderConfig] that specifies whether or
      * not to fetch top sites from the [TopSitesProvider].
      */
     suspend fun getTopSites(
         totalSites: Int,
-        frecencyConfig: FrecencyThresholdOption? = null,
+        frecencyConfig: TopSitesFrecencyConfig? = null,
         providerConfig: TopSitesProviderConfig? = null
     ): List<TopSite>